feat!: Rename serialize module to parameters
Move explicit dump methods to platform
